add setting templates
